Styling list filters
sec. 17, video 174

First, we use content-container to center those filters.

Challenges:
Create input-group partial.
Add selectors for input-group and input-group__item.
Setup flexbox with space below the group.
Setup right margin for the individual items.

We need the <input>, <select> and <DateRangePicker> to stack vertically with margin on the bottom on mobile devices.
For big screens, we want those displayed horizontally with margin on the right.
We will get it done by using media queries.