Styling inputs
sec. 17, video 175

Note that the 3rd party modules, <DateRangePicker>, also use <input>.
If we use generic selector for <input>, we will change the style of that.

Create _inputs.scss.

https://developer.mozilla.org/zh-TW/docs/Web/CSS/box-sizing
Note that by default when the browser is calculating the height or width of something,
it takes the height that we specified. Then it adds on the padding and it adds on the border.
To fix that, we will set the value of box-sizing to border-box in _base.scss.
It's going to set the height of the input to what we want, in our case 50px, then remove the space it needs for the border and the padding.
The rest of the space will be given to the content.

https://sass-lang.com/guide
For styling the <select>, because it will share many of the same styles as the .text-input, we use @extend to inherit the content of .text-input.
What's more, in our case which is not for the teacher's case, the text of the <select> would not be displayed well, we decrease the padding of it.

We also apply the above styling to <ExpenseForm/>.