Styling expense form
sec. 17, video 176

On <AddExpensePage/>, we apply .page-header and .page-header__title to <h1> tag.
Note that the whole point of styling is to create these reusable components that we can use throughout our entire application.
We put the <ExpenseForm/> into a <div> and add .content-container to it.
The reason we use .content-container twice here is that we only want the <h1> and <ExpenseForm/> to be restricted.
If we use .content-container outside of .page-header, the .page-header (gray box) would not go all the way to the edge of the browser.

For the <form>, the 1st thing we want to do is put the error message into the <form> and remove the <div> which originally wrapped the error message and the <form>.
The goal is to display things from top to bottom, so we are going to set display to flex and flex-direction to column (default value is row) to get that done.
To grab everything inside of form which includes our <input> tags and add style to those, we use > to select direct descendants and use selector * for all children.
So right now everything is getting set equal to the full width because we have display: flex in .form.
But that's only going to happen for the direct children of form. If we don't want it to take up the full width, we can toss it in anything, like a <div>.
Now the <div> will be full width but it's children, in this case just the <button>, won't.

We apply the same style of <AddExpensePage/> to <EdditExpensePage/>.
Note that we want to make the "Remove Expense" <button> gray, so we create a modifier of .button, .button--secondary to change the background color.