Styling expense list: part I
sec. 17, video 177

On <ExpenseList/>, we only want to show "Expenses" for mobile devices.
For wider screens, we want to show "Expense" & "Amount".
We can get that done by using @media.
Note that we can subtract a small number to make sure that we never actually have both taking effect at the exact same time.

For the header of the <ExpenseList/>, we will use display: flex and justify-content: space-between to make sure the stuff show up
on the left and right with all of the left over space distributed between the items.