Styling expense list: part II
sec. 17, video 178

For <ExpenseListItem/>, we want to make sure that no matter what we click, we actually do get redirected over to the edit page.
So we put everything into <Link>.

We want "description" and "created" to display on the left (one group) and "amount" on the right (another group).
For mobile devices, we want things going top to bottom. So we use flex-direction: column.
For wider screens, we change the flex-direction to row and set justify-content to space-between.
Note that if we have border on all sides, we're actually going to have 2 px between each item. So we remove the border from the top.

https://www.w3schools.com/css/css3_transitions.asp
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Transitions/Using_CSS_transitions#Using_transitions_when_highlighting_menus

When we move the mouse over the item, we want to change the background color to highlight it. We can use a separate .list-item:hover { … } syntax.
However, we use SCSS feature to target .list-item again. & says take everything from the parent selector, go ahead and toss it on.

https://www.w3schools.com/cssref/css3_pr_word-break.asp

We use work-break: break-all to prevent the description's length overflow.

For the message "No expenses", we create a modifier to align both directions to center and keep the background to none when hovering.
Finally, we put the expense items into a <div> and add margin-bottom to it.