Adding loader
sec. 17, video 179

Originally, we just rendered a <p> before the user's data is ready.
Now we create a <LoadingPage> to render the loader.gif in images folder.
We add some styles to center the image in both directions.
To check the styles of the loading image, we can mark out the authentication process temporarily.
Finally, we create a test case for <LoadingPage>.