Add Babel polyfill
sec. 17, video 180

https://babeljs.io/docs/en/6.26.3/babel-polyfill

Set up and configure Babel polyfill to allow our application to run
in a wider range of browsers and browser versions.
Babel polyfill is going to add in all of those methods (such
Array.includes()) that older browsers don't have.

Installation
$ npm i babel-polyfill@6.26.0

Update webpack.config.js
Now the ES6 and ES7 methods we happened to be using are supported.